1) добавил в класс tensor Ten3D(int size_x, int size_y, int size_z, std::string name, std::vector<std::vector<float>> D); для хард вектора

+++ Tensor_Class/Neuro2.h
@@ -12,7 +12,7 @@ public:
 private:
 	std::vector < std::vector<float >> vector_Layers; //вектор слоев
 	std::vector < std::vector<float >> w;	          //матрица весов
-	std::vector <float> layerSoftMax;
+	std::vector <int> layerSoftMax;
 	void Layer();
 	int size_in_layer;//потом понадобится вроде
 	float func(float x);// ф-ция активации
